add noised laplace to sample process

+++ src/document.h
@@ -48,7 +48,7 @@ namespace multiverso { namespace lightlda
         int32_t& cursor_;
 
         //文档中的词的干扰词，以及干扰词的权重。按照位置存储
-        std::vector<std::vector<std::pair<int32_t, double>>> noise_words;
+        std::vector<std::vector<std::pair<int32_t, float>>> noise_words;
 
         // No copying allowed
         Document(const Document&);
